refactor language server tests
