+++ .vscode/launch.json
@@ -98,6 +98,7 @@
             "runtimeExecutable": "${execPath}",
             "args": [
                 "${workspaceFolder}/src/test",
+                "--disable-extensions",
                 "--extensionDevelopmentPath=${workspaceFolder}",
                 "--extensionTestsPath=${workspaceFolder}/out/test"
             ],
