+++ src/test/languageServers/jedi/autocomplete/pep484.test.ts
@@ -1,23 +1,23 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+'use strict';
+
 import * as assert from 'assert';
 import * as path from 'path';
 import * as vscode from 'vscode';
-import { rootWorkspaceUri } from '../common';
-import { IsLanguageServerTest } from '../constants';
-import { closeActiveWindows, initialize, initializeTest } from '../initialize';
-import { UnitTestIocContainer } from '../unittests/serviceRegistry';
+import { EXTENSION_ROOT_DIR } from '../../../../client/common/constants';
+import { rootWorkspaceUri } from '../../../common';
+import { closeActiveWindows, initialize, initializeTest } from '../../../initialize';
+import { UnitTestIocContainer } from '../../../unittests/serviceRegistry';
 
-const autoCompPath = path.join(__dirname, '..', '..', '..', 'src', 'test', 'pythonFiles', 'autocomp');
+const autoCompPath = path.join(EXTENSION_ROOT_DIR, 'src', 'test', 'pythonFiles', 'autocomp');
 const filePep484 = path.join(autoCompPath, 'pep484.py');
 
 suite('Autocomplete PEP 484', () => {
     let isPython2: boolean;
     let ioc: UnitTestIocContainer;
     suiteSetup(async function () {
-        // https://github.com/Microsoft/PTVS/issues/3917
-        if (IsLanguageServerTest()) {
-            // tslint:disable-next-line:no-invalid-this
-            this.skip();
-        }
         await initialize();
         initializeDI();
         isPython2 = await ioc.getPythonMajorVersion(rootWorkspaceUri) === 2;
