+++ src/test/languageServers/jedi/symbolProvider.unit.test.ts
@@ -13,13 +13,13 @@ import {
     TextDocument, Uri
 } from 'vscode';
 import { LanguageClient } from 'vscode-languageclient';
-import { IFileSystem } from '../../client/common/platform/types';
-import { splitParent } from '../../client/common/utils/string';
-import { parseRange } from '../../client/common/utils/text';
-import { IServiceContainer } from '../../client/ioc/types';
-import { JediFactory } from '../../client/languageServices/jediProxyFactory';
-import { IDefinition, ISymbolResult, JediProxyHandler } from '../../client/providers/jediProxy';
-import { JediSymbolProvider, LanguageServerSymbolProvider } from '../../client/providers/symbolProvider';
+import { IFileSystem } from '../../../client/common/platform/types';
+import { splitParent } from '../../../client/common/utils/string';
+import { parseRange } from '../../../client/common/utils/text';
+import { IServiceContainer } from '../../../client/ioc/types';
+import { JediFactory } from '../../../client/languageServices/jediProxyFactory';
+import { IDefinition, ISymbolResult, JediProxyHandler } from '../../../client/providers/jediProxy';
+import { JediSymbolProvider, LanguageServerSymbolProvider } from '../../../client/providers/symbolProvider';
 
 const assertArrays = require('chai-arrays');
 use(assertArrays);
